feat: Split encoding pipeline: diskreader pkg (unified with diskwriter pkg)
